+++ packages/client-ledger/src/sdk/sdk.ts
@@ -91,6 +91,7 @@ export class LedgerClient extends ClientSDK {
                 path: path$,
                 headers: headers$,
                 query: query$,
+                timeoutMs: options?.timeoutMs || this.options$.timeoutMs || -1,
             },
             options
         );