+++ packages/client-ledger/RELEASES.md
@@ -47,3 +47,13 @@ Based on:
 - [typescript v1.0.0-alpha.4] packages/client-ledger
 ### Releases
 - [NPM v1.0.0-alpha.4] https://www.npmjs.com/package/@formance/sdk-ledger/v/1.0.0-alpha.4 - packages/client-ledger
+
+## 2024-07-12 00:15:29
+### Changes
+Based on:
+- OpenAPI Doc  
+- Speakeasy CLI 1.332.1 (2.366.1) https://github.com/speakeasy-api/speakeasy
+### Generated
+- [typescript v1.0.0-alpha.5] packages/client-ledger
+### Releases
+- [NPM v1.0.0-alpha.5] https://www.npmjs.com/package/@formance/sdk-ledger/v/1.0.0-alpha.5 - packages/client-ledger